Fix vault toke TTL handling
In case the authentication token expire, _all_ created leased which was
created in that session, are expired as well.
From [1] you can found an explanation:
```
Revocation can happen manually via the API, via the vault revoke cli command,
or automatically by Vault. When a lease is expired, Vault will automatically
revoke that lease. When a token is revoked, Vault will revoke all leases that
were created using that token.
```

So we use [2] to increase our token expire time and also make sure that
we use only one token per Key.
As there is a other concept of "Explicit maximum lifetime" [3]. If I
understood this concept right, there is no way to have an unlimited token.

[1] https://www.vaultproject.io/docs/concepts/lease.html
[2] https://www.vaultproject.io/api/auth/token/index.html#renew-a-token
[3] https://www.vaultproject.io/docs/concepts/tokens.html#explicit-max-ttls