Add automatic deletion of Instances
If an instance has a tag called openqa_ttl the number is used in seconds
to set the max age (time to life). If this ttl is reached, we delete the
instance automatically.

A instance in state DELETING still produce a email notification. As it
might happen that a delete_instance() just went wrong somehow on the CSP
site and manual interaction is needed.