Add Dockerfile and make pcw standalone
This commit improves pcw to be able to run as a standalone container
that only requires a correspodning `pcw.ini` file.

It adds a Dockerfile for building pcw and corresponding changes to pcw
itself, so it works as standalone container. For that, the
ALLOWED_HOSTS is set to accept any incoming connection from any
host.

Further it improves the documentation to match the current state of
Publiccloud-Watcher.